Add PR size validator
Adds a new 'size' validator that allows for validating the size of PRs
in terms of changes. Currently, this only lets you validate that the
total no. of changed lines is below a maximum using the 'max' option.

It also supports an 'ignore' setting to allow excluding certain files
from the total size (e.g. for ignoring automatically generated files
that increase the size a lot).

Things it does not support but could be added in future:
  - Using different options than 'max'
  - Validating additions and deletions separately (e.g allowing 300
    additions max, 600 deletions max)
  - Including entirely deleted files in the total (can very easily be
    added)

Other changes that were needed to get this working:
  - Update 'max' option to support integers as well as Arrays