chore(package): Update probot, support for base64-encoded keys, etc. (#200)
Update the 'probot' npm package to the latest version.

This requires a number of small changes to deprecated methods in the github
context API. A couple of these were breaking, but most were just
deprecation warnings. It also requires a bump to `probot-scheduler`.

The main motivation for this was to fix a bug where the PRIVATE_KEY
environment variable was not being correctly recognised if it was base64
encoded, thus preventing deployment of the bot. base64-encoded keys are
supported in later versions of the probot lib:

https://probot.github.io/docs/configuration/

Support for this was added in 7.1.0:

https://github.com/probot/probot/pull/624