Fix FetchConfigFile from forked PR
FetchConfigFile would fail when trying to fetch a modified config from a
PR living on a fork. It would try to use the pull_request's ref (aka the
branch name) on the original repository, which would fail since said
branch might not exist.

We instead use the sha as a ref, which works much better.