Add support for default config files
Adds support for specifying default config files in a central repository.
This allows for providing default settings for Mergeable that can be
overriden in individual repos. The purpose is to make it easier to manage
Mergeable across a large number of repos.

The behaviour is quite basic, so there is no support for overriding individual
settings in the configs. Mergeable will use either the default file or the
file in the repo/PR, but never both.

The approach is based on the Probot library which when loading the
config file will look in a repo named '.github' as a fallback. Because
Mergeable can only be granted permission to a specific file path, the
default file needs a somewhat confusing path
('<org>/.github/.github/mergeable.yml').