Add an option to use PR title as commit message
Added a merge setting option called `use_pr_as_commit_message` that uses the PR title and body as the commit message. This allows users to set rules requiring certain commit messages without forcing the user to squash and `git push --force`. If we want the commit message to always contain the ticket, just force the user to update the PR title instead of pushing again.